+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/interfaces/Action.kt
@@ -2,6 +2,7 @@ package eu.vendeli.tgbot.interfaces
 
 import eu.vendeli.tgbot.TelegramBot
 import eu.vendeli.tgbot.types.User
+import eu.vendeli.tgbot.types.chat.Chat
 import eu.vendeli.tgbot.types.internal.Response
 import eu.vendeli.tgbot.utils.makeRequestAsync
 import eu.vendeli.tgbot.utils.makeSilentRequest
@@ -24,24 +25,17 @@ abstract class Action<ReturnType> : TgAction<ReturnType>() {
         via.makeSilentRequest(method, parameters)
     }
 
-    /**
-     * Make a request for action.
-     *
-     * @param to Recipient
-     * @param via Instance of the bot through which the request will be made.
-     */
     open suspend fun send(to: Long, via: TelegramBot) {
         parameters["chat_id"] = to
         via.makeSilentRequest(method, parameters)
     }
 
-    /**
-     * Make a request for action.
-     *
-     * @param to Recipient
-     * @param via Instance of the bot through which the request will be made.
-     */
     open suspend fun send(to: User, via: TelegramBot) {
+        parameters["chat_id"] = to.id
+        via.makeSilentRequest(method, parameters)
+    }
+
+    open suspend fun send(to: Chat, via: TelegramBot) {
         parameters["chat_id"] = to.id
         via.makeSilentRequest(method, parameters)
     }
@@ -60,12 +54,14 @@ abstract class Action<ReturnType> : TgAction<ReturnType>() {
         return via.makeRequestAsync(method, parameters, returnType, wrappedDataType)
     }
 
-    /**
-     * Make a request for action returning its [Response].
-     *
-     * @param to Recipient
-     * @param via Instance of the bot through which the request will be made.
-     */
+    open suspend fun sendAsync(
+        to: String,
+        via: TelegramBot,
+    ): Deferred<Response<out ReturnType>> {
+        parameters["chat_id"] = to
+        return via.makeRequestAsync(method, parameters, returnType, wrappedDataType)
+    }
+
     open suspend fun sendAsync(
         to: User,
         via: TelegramBot,
@@ -74,17 +70,11 @@ abstract class Action<ReturnType> : TgAction<ReturnType>() {
         return via.makeRequestAsync(method, parameters, returnType, wrappedDataType)
     }
 
-    /**
-     * Make a request for action returning its [Response].
-     *
-     * @param to Recipient
-     * @param via Instance of the bot through which the request will be made.
-     */
     open suspend fun sendAsync(
-        to: String,
+        to: Chat,
         via: TelegramBot,
     ): Deferred<Response<out ReturnType>> {
-        parameters["chat_id"] = to
+        parameters["chat_id"] = to.id
         return via.makeRequestAsync(method, parameters, returnType, wrappedDataType)
     }
 }