+++ telegram-bot/src/test/kotlin/eu/vendeli/api/botactions/BotMainActionsTest.kt
@@ -7,20 +7,20 @@ import eu.vendeli.tgbot.api.botactions.getWebhookInfo
 import eu.vendeli.tgbot.api.botactions.setWebhook
 import eu.vendeli.tgbot.types.internal.foldResponse
 import io.kotest.matchers.nulls.shouldNotBeNull
-import io.kotest.matchers.shouldBe
 import io.kotest.matchers.string.shouldBeEmpty
+import io.kotest.matchers.string.shouldStartWith
 
 class BotMainActionsTest : BotTestContext() {
     @Test
     suspend fun `get updates method testing`() {
-        val result = getUpdates().sendAsync(bot).await().shouldSuccess()
+        val result = getUpdates().sendAsync(bot).shouldSuccess()
 
         result.shouldNotBeNull()
     }
 
     @Test
     suspend fun `get webhook info method testing`() {
-        val result = getWebhookInfo().sendAsync(bot).await().shouldSuccess()
+        val result = getWebhookInfo().sendAsync(bot).shouldSuccess()
 
         result.shouldNotBeNull()
         result.url.shouldBeEmpty()
@@ -31,10 +31,10 @@ class BotMainActionsTest : BotTestContext() {
         setWebhook("https://vendeli.eu").send(bot)
         setWebhook("https://vendeli.eu/1").sendAsync(bot)
             .foldResponse({ println(result) }, { println(errorCode) })
-        val result = getWebhookInfo().sendAsync(bot).await().shouldSuccess()
+        val result = getWebhookInfo().sendAsync(bot).shouldSuccess()
 
         result.shouldNotBeNull()
-        result.url shouldBe "https://vendeli.eu"
+        result.url shouldStartWith "https://vendeli.eu"
 
         deleteWebhook().send(bot)
     }