+++ telegram-bot/src/test/kotlin/eu/vendeli/api/botactions/CreateInvoiceLinkTest.kt
@@ -21,7 +21,7 @@ class CreateInvoiceLinkTest : BotTestContext() {
             LabeledPrice("test3", 1000),
         ) {
             "test2"
-        }.sendAsync(bot).await().shouldSuccess()
+        }.sendAsync(bot).shouldSuccess()
 
         result.shouldNotBeNull()
         result shouldContain "https://t.me"